+++ app/src/main/java/com/tomclaw/appsend_rb/util/AppIconData.kt
@@ -3,11 +3,11 @@ package com.tomclaw.appsend_rb.util
 import android.os.Parcel
 import android.os.Parcelable
 
-class AppIconData(val packageName: String, val versionCode: Int) : Parcelable {
+class AppIconData(val packageName: String, val versionCode: Long) : Parcelable {
 
     override fun writeToParcel(dest: Parcel, flags: Int) = with(dest) {
         writeString(packageName)
-        writeInt(versionCode)
+        writeLong(versionCode)
     }
 
     override fun describeContents(): Int = 0
@@ -15,7 +15,7 @@ class AppIconData(val packageName: String, val versionCode: Int) : Parcelable {
     companion object CREATOR : Parcelable.Creator<AppIconData> {
         override fun createFromParcel(parcel: Parcel): AppIconData {
             val packageName = parcel.readString().orEmpty()
-            val versionCode = parcel.readInt()
+            val versionCode = parcel.readLong()
             return AppIconData(packageName, versionCode)
         }
